requirements.txt
# Ops Voice Co-Pilot — GCP backend (monolith + microservices)
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
pydantic-settings>=2.0.0,<3.0.0
google-genai>=1.0.0
google-cloud-logging>=3.0.0,<4.0.0
websockets>=12.0,<14.0
httpx>=0.28.1,<1.0.0
